refactor(gateway): extract GatewayPlatformMixin from run.py (slice 25 of #54962)
